working celery config
