+++ app.py
@@ -4,6 +4,7 @@ import os
 # https://realpython.com/flask-connexion-rest-api/
 # https://github.com/realpython/materials/tree/master/flask-connexion-rest-part-4
 from celery import Celery
+from tasks import add_together
 
 import config
 
@@ -21,23 +22,10 @@ connex_app.app.logger.setLevel(gunicorn_logger.level)
 # Read the openapi.yml file to configure the endpoints
 connex_app.add_api('openapi.yml', strict_validation=False)
 
-def make_celery(app):
-    celery = Celery(
-        app.import_name,
-        backend=app.config['CELERY_RESULT_BACKEND'],
-        broker=app.config['CELERY_BROKER_URL']
-    )
-    celery.conf.update(app.config)
+# celery = make_celery(connex_app.app)
 
-    class ContextTask(celery.Task):
-        def __call__(self, *args, **kwargs):
-            with app.app_context():
-                return self.run(*args, **kwargs)
-
-    celery.Task = ContextTask
-    return celery
-
-celery = make_celery(connex_app.app)
+result = add_together.delay(23, 42)
+result.wait()
 
 if __name__ == "__main__":
     connex_app.run(port=3100, debug=os.environ['DEBUG'])
