+++ config.py
@@ -13,9 +13,6 @@ app = connex_app.app
 
 ma = Marshmallow(app)
 
-app.config['CELERY_BROKER_URL'] ='redis://localhost:6379'
-app.config['CELERY_RESULT_BACKEND'] ='redis://localhost:6379'
-
 setDefaultGoogleCloudProject()
