+++ tasks.py
@@ -1,0 +1,32 @@
+import os
+
+from celery import Celery
+from flask import Flask
+
+flask_app = Flask(__name__)
+flask_app.config.update(
+    CELERY_BROKER_URL=os.environ['CELERY_BROKER_URL'],
+    CELERY_RESULT_BACKEND=os.environ['CELERY_RESULT_BACKEND']
+)
+
+def make_celery(app):
+    celery = Celery(
+        app.import_name,
+        backend=app.config['CELERY_RESULT_BACKEND'],
+        broker=app.config['CELERY_BROKER_URL']
+    )
+    # celery.conf.update(app.config)
+
+    class ContextTask(celery.Task):
+        def __call__(self, *args, **kwargs):
+            with app.app_context():
+                return self.run(*args, **kwargs)
+
+    celery.Task = ContextTask
+    return celery
+
+celery = make_celery(flask_app)
+
+@celery.task()
+def add_together(a, b):
+    return a + b
